5/25までの変更点 次回はGrowthの処理を書く

+++ Assets/Scripts/Main/Enemy/BaseEnemy.cs
@@ -2,9 +2,9 @@ using System;
 using UnityEngine;
 
 internal class BaseEnemy : MonoBehaviour {
-    public static event Action<int> OnDefeated;
+    public static event Action<float> OnDefeated;
 
-    [SerializeField] private int _baseExp = 10;
+    [SerializeField] private float _baseExp = 10;
     [SerializeField] protected int _hp = 20;
 
     protected void Defeated() {
